fix: renamed features on redux
