+++ src/pages/login/userpassword-modal.tsx
@@ -2,8 +2,8 @@ import { Link } from "react-router-dom";
 import { BackButton } from "../../components/micro-components/back-button";
 import { MainButton } from "../../components/micro-components/main-button";
 import { useDispatch, useSelector } from "react-redux";
-import { setUserLogin } from "../../features/user-login-slice";
-import { setAuthUser } from "../../features/get-auth-user-slice";
+import { setUserLogin } from "../../features/login-slice";
+import { setAuthUser } from "../../features/auth-user-slice";
 import { RootState } from "../../store";
 import { api } from "../../lib/axios";
