+++ src/store.ts
@@ -1,9 +1,9 @@
 import { configureStore } from "@reduxjs/toolkit";
-import userReducer from '../src/features/get-user-input-slice' 
-import validationPasswordReducer from '../src/features/validation-password-slice'
-import userAddressReducer from '../src/features/get-user-address-slice';
-import userLoginReducer from '../src/features/user-login-slice';
-import authUserReducer from '../src/features/get-auth-user-slice'
+import userReducer from './features/register-data-user' 
+import validationPasswordReducer from './features/register-user-password'
+import userAddressReducer from './features/register-user-address';
+import userLoginReducer from './features/login-slice';
+import authUserReducer from './features/auth-user-slice'
 
 export const store = configureStore({
     reducer:{
